requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
slowapi
python-jose
passlib